requirements.txt
streamlit
dspy-ai
python-dotenv
openai